requirements.txt
Adafruit-Blinka==8.66.2
Adafruit-Blinka-Raspberry-Pi5-Neopixel==1.0.0rc2
adafruit-circuitpython-busdevice==5.2.13
adafruit-circuitpython-connectionmanager==3.1.5
adafruit-circuitpython-dht==4.0.9
adafruit-circuitpython-requests==4.1.13
adafruit-circuitpython-typing==1.12.2
Adafruit-PlatformDetect==3.84.0
Adafruit-PureIO==1.1.11
binho-host-adapter==0.1.6
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
colorzero==2.0
contourpy==1.3.3
cycler==0.12.1
Flask==3.1.2
Flask-Mail==0.10.0
fonttools==4.60.1
gpiozero==2.0.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.9
lgpio==0.2.2.0
MarkupSafe==3.0.3
matplotlib==3.10.6
numpy==2.3.3
packaging==25.0
pandas==2.3.3
pillow==11.3.0
pyftdi==0.57.1
pyparsing==3.2.5
pyserial==3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pyusb==1.3.1
requests==2.32.5
rpi-ws281x==5.0.0
RPi.GPIO==0.7.1
six==1.17.0
sysv-ipc==1.1.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
